# Conflicts:
#	commons/src/main/java/commons/Card.java

+++ server/src/main/java/server/services/CardService.java
@@ -75,7 +75,11 @@ public class CardService {
      * @param id
      */
     public void deleteCardById(Long id) {
-        cardRepository.deleteById(id);
+        Card card = cardRepository.getById(id);
+        if(card != null) {
+            cardRepository.getById(id).removeCardsFromTags();
+            cardRepository.deleteById(id);
+        }
     }
 
     /**
